Split out themeActivated
